+++ src/main/java/org/truffleruby/language/methods/DeclarationContext.java
@@ -119,6 +119,25 @@ public class DeclarationContext {
         return RubyArguments.getDeclarationContext(visibilityFrame).visibility;
     }
 
+    /** See rb_vm_cref_in_context() in CRuby */
+    public static Visibility findVisibilityCheckSelfAndDefaultDefinee(RubyModule module, Frame callerFrame) {
+        final Frame visibilityFrame = lookupVisibility(callerFrame);
+        final DeclarationContext declarationContext = RubyArguments.getDeclarationContext(visibilityFrame);
+        if (declarationContext == DeclarationContext.NONE) {
+            // For Java core methods, e.g. method(:attr_accessor).to_proc as in spec/truffle/always_inlined_spec.rb
+            // The generated Proc uses the DeclarationContext from Module#attr_accessor and knows nothing about the
+            // lexical scope surrounding the Proc#call.
+            // Maybe we should go to the 2nd caller in this case, skipping the SetReceiverNode CallTarget.
+            return Visibility.PUBLIC;
+        } else if (RubyArguments.getSelf(callerFrame) != module) {
+            return Visibility.PUBLIC;
+        } else if (declarationContext.getModuleToDefineMethods() != module) {
+            return Visibility.PUBLIC;
+        } else {
+            return declarationContext.visibility;
+        }
+    }
+
     private static void changeVisibility(Frame frame, Visibility newVisibility) {
         final Frame visibilityFrame = lookupVisibility(frame);
         final DeclarationContext oldDeclarationContext = RubyArguments.getDeclarationContext(visibilityFrame);