+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Compatibility:
 * Include the type name in exception messages from `rb_check_type` (#2307).
 * Fix `Hash#rehash` to remove duplicate keys after modifications (#2266, @MattAlp)
 * Only fail `rb_check_type` for typed data, not wrapped untyped structs (#2331).
+* Decide the visibility in `Module#define_method` based on `self` and the default definee (#2334).
 
 Performance: